Moving bundle classes to util package
