+++ vraptor-core/src/main/java/br/com/caelum/vraptor/util/SafeResourceBundle.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package br.com.caelum.vraptor.core;
+package br.com.caelum.vraptor.util;
 
 import static com.google.common.base.Preconditions.checkNotNull;
